Update mmu.c for TLB

+++ week09/mmu.c
@@ -3,7 +3,6 @@
 #include <signal.h>
 #include <stdlib.h>
 #include <string.h>
-#include <string.h>
 #include <stdbool.h>
 #include <fcntl.h>
 #include <sys/mman.h>
@@ -18,8 +17,16 @@ struct PTE {
     int counter;
 };
 
+struct TLBEntry {
+    int page;
+    int frame;
+};
+
 int pager_pid;
 struct PTE *pagetable;
+struct TLBEntry *tlb;
+int tlb_index = 0;
+int tlb_size = 0;
 
 int number_of_pages;
 char *str;
@@ -39,23 +46,42 @@ void process_memory_access(const char *access) {
         exit(1);
     }
 
-    if (!pagetable[page].valid) {
-        printf("It is not a valid page --> page fault\n");
-        printf("Ask pager to load it from disk (SIGUSR1 signal) and wait\n");
+    // Check the TLB for the requested page
+    int frame = -1;
+    for (int i = 0; i < tlb_size; i++) {
+        if (tlb[i].page == page) {
+            frame = tlb[i].frame;
+            break;
+        }
+    }
 
-        pagetable[page].referenced = getpid();
-        kill(pager_pid, SIGCONT);
-
-        pause();
-    } else {
-        printf("Page is valid (hit)\n");
+    if (frame != -1) {
+        printf("Page is valid (hit) in TLB\n");
         hits++;
+    } else {
+        printf("Page is not in TLB --> TLB miss\n");
+        if (!pagetable[page].valid) {
+            printf("It is not a valid page --> page fault\n");
+            printf("Ask pager to load it from disk (SIGUSR1 signal) and wait\n");
+
+            pagetable[page].referenced = getpid();
+            kill(pager_pid, SIGCONT);
+
+            pause();
+        } else {
+            printf("Page is valid (hit) in page table\n");
+            hits++;
+        }
+
+        // Update TLB with the new mapping
+        tlb[tlb_index].page = page;
+        tlb[tlb_index].frame = pagetable[page].frame;
+        tlb_index = (tlb_index + 1) % tlb_size;
     }
 
     accesses_count++;
-
-
     printf("MMU resumed by SIGCONT signal from pager\n");
+
     if (mode == 'W') pagetable[page].dirty = true;
 }
 
@@ -68,18 +94,25 @@ void sigcont_handler(int signo) {
 }
 
 void initialize_page_table(int num_pages) {
-    int fd = open("/tmp/ex2/pagetable", 02);
+    int fd = open("/tmp/ex2/pagetable", O_RDWR);
     if (fd == -1) {
         perror("open");
         exit(1);
     }
 
     ftruncate(fd, num_pages * sizeof(struct PTE) * 1000);
-    pagetable = mmap(NULL, num_pages * sizeof(struct PTE) * 1000, 0x1 | 0x2, 0x01, fd, 0);
+    pagetable = mmap(NULL, num_pages * sizeof(struct PTE) * 1000, PROT_READ | PROT_WRITE, MAP_SHARED, fd, 0);
+    tlb_size = num_pages * 20 / 100;
+    if (tlb_size <= 0)tlb_size = 1;
+    tlb = malloc(tlb_size * sizeof(struct TLBEntry));
+    for (int i = 0; i < tlb_size; i++) {
+        tlb[i].page = -1;
+        tlb[i].frame = -1;
+    }
     if (pagetable == MAP_FAILED) {
         perror("mmap");
         close(fd);
-        exit(EXIT_FAILURE);
+        exit(1);
     }
 
     printf("-------------------------\n");
@@ -97,11 +130,11 @@ void cleanup() {
 
     kill(pager_pid, SIGUSR1);
 
-    double hit_ratio = (double)hits / accesses_count;
-    printf("Hit Ratio: %.4f\n", hit_ratio);
+    double tlb_miss_ratio = 1.0 - (double) hits / accesses_count;
+    printf("TLB Miss Ratio: %.4f\n", tlb_miss_ratio);
 
     printf("MMU terminates.\n");
-    exit(EXIT_SUCCESS);
+    exit(0);
 }
 
 int main(int argc, char *argv[]) {
@@ -121,13 +154,15 @@ int main(int argc, char *argv[]) {
     signal(SIGUSR1, sigusr1_handler);
     signal(SIGCONT, sigcont_handler);
 
+
     char *token = strtok(str, " ");
     while (token != NULL) {
         process_memory_access(token);
         printf("Page table\n");
         for (int i = 0; i < num_pages; i++)
-            printf("Page %d ---> valid=%d, frame=%d, dirty=%d, referenced=%d, referenced counter=%d\n",
-                   i, pagetable[i].valid, pagetable[i].frame, pagetable[i].dirty, pagetable[i].referenced, pagetable[i].counter);
+            printf("Page %d ---> valid=%d, frame=%d, dirty=%d, referenced=%d, counter=%d\n",
+                   i, pagetable[i].valid, pagetable[i].frame, pagetable[i].dirty, pagetable[i].referenced,
+                   pagetable[i].counter);
 
         token = strtok(NULL, " ");
     }
